split original server/server.go to server/server.go, sotw/server.go and rest/server.go
